complete

+++ src/main/scala/me/gladwell/kata/checkout/Offer.scala
@@ -1,3 +1,3 @@
 package me.gladwell.kata.checkout
 
-case class Offer(sku: String, products: Seq[ShoppingItem], discount: Double)
+case class Offer(sku: String, products: Seq[ShoppingItem], discount: BigDecimal)
