+++ src/main/scala/me/gladwell/kata/checkout/ShoppingItem.scala
@@ -1,3 +1,3 @@
 package me.gladwell.kata.checkout
 
-case class ShoppingItem(sku: String, price: Double)
+case class ShoppingItem(sku: String, price: BigDecimal)
